+++ vm/src/vm/hooks.rs
@@ -26,7 +26,7 @@ type BeforeFirstStepHookFunc = Arc<
     dyn Fn(
             &mut VirtualMachine,
             &mut CairoRunner,
-            &HashMap<usize, Vec<Box<dyn Any>>>,
+            &HashMap<u64, Vec<Box<dyn Any>>>,
         ) -> Result<(), VirtualMachineError>
         + Sync
         + Send,
@@ -37,7 +37,7 @@ type StepHookFunc = Arc<
             &mut VirtualMachine,
             &mut dyn HintProcessor,
             &mut ExecutionScopes,
-            &HashMap<usize, Vec<Box<dyn Any>>>,
+            &HashMap<u64, Vec<Box<dyn Any>>>,
             &HashMap<String, Felt252>,
         ) -> Result<(), VirtualMachineError>
         + Sync
@@ -72,7 +72,7 @@ impl VirtualMachine {
     pub fn execute_before_first_step(
         &mut self,
         runner: &mut CairoRunner,
-        hint_data_dictionary: &HashMap<usize, Vec<Box<dyn Any>>>,
+        hint_data_dictionary: &HashMap<u64, Vec<Box<dyn Any>>>,
     ) -> Result<(), VirtualMachineError> {
         if let Some(hook_func) = self.hooks.clone().before_first_step {
             (hook_func)(self, runner, hint_data_dictionary)?;
@@ -85,7 +85,7 @@ impl VirtualMachine {
         &mut self,
         hint_executor: &mut dyn HintProcessor,
         exec_scope: &mut ExecutionScopes,
-        hint_data_dictionary: &HashMap<usize, Vec<Box<dyn Any>>>,
+        hint_data_dictionary: &HashMap<u64, Vec<Box<dyn Any>>>,
         constants: &HashMap<String, Felt252>,
     ) -> Result<(), VirtualMachineError> {
         if let Some(hook_func) = self.hooks.clone().pre_step_instruction {
@@ -105,7 +105,7 @@ impl VirtualMachine {
         &mut self,
         hint_executor: &mut dyn HintProcessor,
         exec_scope: &mut ExecutionScopes,
-        hint_data_dictionary: &HashMap<usize, Vec<Box<dyn Any>>>,
+        hint_data_dictionary: &HashMap<u64, Vec<Box<dyn Any>>>,
         constants: &HashMap<String, Felt252>,
     ) -> Result<(), VirtualMachineError> {
         if let Some(hook_func) = self.hooks.clone().post_step_instruction {
@@ -166,7 +166,7 @@ mod tests {
         fn before_first_step_hook(
             _vm: &mut VirtualMachine,
             _runner: &mut CairoRunner,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
         ) -> Result<(), VirtualMachineError> {
             Err(VirtualMachineError::Unexpected)
         }
@@ -175,7 +175,7 @@ mod tests {
             _vm: &mut VirtualMachine,
             _hint_processor: &mut dyn HintProcessor,
             _exec_scope: &mut ExecutionScopes,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
             _constants: &HashMap<String, Felt252>,
         ) -> Result<(), VirtualMachineError> {
             Err(VirtualMachineError::Unexpected)
@@ -185,7 +185,7 @@ mod tests {
             _vm: &mut VirtualMachine,
             _hint_processor: &mut dyn HintProcessor,
             _exec_scope: &mut ExecutionScopes,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
             _constants: &HashMap<String, Felt252>,
         ) -> Result<(), VirtualMachineError> {
             Err(VirtualMachineError::Unexpected)
@@ -251,7 +251,7 @@ mod tests {
         fn before_first_step_hook(
             _vm: &mut VirtualMachine,
             _runner: &mut CairoRunner,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
         ) -> Result<(), VirtualMachineError> {
             Ok(())
         }
@@ -260,7 +260,7 @@ mod tests {
             _vm: &mut VirtualMachine,
             _hint_processor: &mut dyn HintProcessor,
             _exec_scope: &mut ExecutionScopes,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
             _constants: &HashMap<String, Felt252>,
         ) -> Result<(), VirtualMachineError> {
             Ok(())
@@ -270,7 +270,7 @@ mod tests {
             _vm: &mut VirtualMachine,
             _hint_processor: &mut dyn HintProcessor,
             _exec_scope: &mut ExecutionScopes,
-            _hint_data: &HashMap<usize, Vec<Box<dyn Any>>>,
+            _hint_data: &HashMap<u64, Vec<Box<dyn Any>>>,
             _constants: &HashMap<String, Felt252>,
         ) -> Result<(), VirtualMachineError> {
             Ok(())
