+++ vm/src/vm/runners/builtin_runner/bitwise.rs
@@ -68,11 +68,14 @@ impl BitwiseBuiltinRunner {
         address: Relocatable,
         memory: &Memory,
     ) -> Result<Option<MaybeRelocatable>, RunnerError> {
-        let index = address.offset % self.cells_per_instance as usize;
+        let index = address.offset % self.cells_per_instance as u64;
         if index <= 1 {
             return Ok(None);
         }
-        let x_addr = Relocatable::from((address.segment_index, address.offset - index));
+        let x_addr = Relocatable::from((
+            address.segment_index as isize,
+            (address.offset - index) as usize,
+        ));
         let y_addr = (x_addr + 1_usize)?;
 
         let num_x = memory.get(&x_addr);
@@ -146,7 +149,7 @@ impl BitwiseBuiltinRunner {
                 .memory
                 .get_relocatable(stop_pointer_addr)
                 .map_err(|_| RunnerError::NoStopPointer(Box::new(BITWISE_BUILTIN_NAME)))?;
-            if self.base as isize != stop_pointer.segment_index {
+            if self.base as isize != stop_pointer.segment_index as isize {
                 return Err(RunnerError::InvalidStopPointerIndex(Box::new((
                     BITWISE_BUILTIN_NAME,
                     stop_pointer,
@@ -156,14 +159,14 @@ impl BitwiseBuiltinRunner {
             let stop_ptr = stop_pointer.offset;
             let num_instances = self.get_used_instances(segments)?;
             let used = num_instances * self.cells_per_instance as usize;
-            if stop_ptr != used {
+            if stop_ptr != used as u64 {
                 return Err(RunnerError::InvalidStopPointer(Box::new((
                     BITWISE_BUILTIN_NAME,
                     Relocatable::from((self.base as isize, used)),
-                    Relocatable::from((self.base as isize, stop_ptr)),
+                    Relocatable::from((self.base as isize, stop_ptr as usize)),
                 ))));
             }
-            self.stop_ptr = Some(stop_ptr);
+            self.stop_ptr = Some(stop_ptr as usize);
             Ok(stop_pointer_addr)
         } else {
             let stop_ptr = self.base;
